requirements.txt
lark-oapi>=1.5.0
PyYAML>=6.0
